Extract EquiJoinPlanner from MongoDbPlanner into separate file
